GB3-1715: Move the listener to search-window.component.ts
