+++ src/app/shared/directives/search-result-identifier.directive.ts
@@ -6,6 +6,7 @@ import {Directive, ElementRef, EventEmitter, Input, Output} from '@angular/core'
 })
 export class SearchResultIdentifierDirective {
   @Input() public isMapResult: boolean = false;
+  @Input() public text: string = '';
   @Output() public readonly addResultFromArrowNavigation = new EventEmitter<void>();
   @Output() public readonly removeResultFromArrowNavigation = new EventEmitter<void>();
   constructor(public readonly host: ElementRef<HTMLElement>) {}
